Centralise mocked 'renderData' and 'getRenderData' into a helper file.
